+++ testapps/GHIssues/src/main/java/ghissues/gh820/Gh820IntService.java
@@ -1,0 +1,8 @@
+package ghissues.gh820;
+
+public class Gh820IntService implements Gh820Service<Integer> {
+    @Override
+    public Integer get() {
+        return 123;
+    }
+}