+++ testapps/GHIssues/src/main/java/ghissues/gh820/Gh820Service.java
@@ -1,0 +1,5 @@
+package ghissues.gh820;
+
+public interface Gh820Service<ENTITY_TYPE> {
+    ENTITY_TYPE get();
+}